# Speed up time_index_to_datetime()